FTP is in passive mode now
I had to convert FTP to be in passive mode instead of active
mode because for whatever reason the Go and FTP are not opening
sockets correctly in active mode. I searched everywhere but I
can't seem to find a way to fix it.

The data connection is now being initiated and transferring data
correctly. I have to refactor the code to account for the new
way that this works. I also need to actually implement the real
ls command.