Files are now transferred successfully!
I created a way for the files to be sent through
the data connection.

In order for the file send to actually work I had
to create a TYPE handler and set it to always return
success. This lets you shut the FTP client up when
it starts complaining about bare line feeds and ascii
mode by setting the type to bin for binary.